feat: complete Week 5-7 Metacognitive Evolution for v0.6.0
Metacognitive Self-Improvement System:
- Enhanced PerformanceMonitor with degradation detection
  * Success rate, quality, and latency tracking
  * Bottleneck identification (high latency, low quality, failures)
  * Trend analysis with linear regression (R-squared confidence)
  * 95th percentile latency tracking
  * Severity classification (None/Minor/Moderate/Severe)

- Enhanced WeaknessAnalyzer with pattern intelligence
  * LLM-powered failure pattern identification
  * Pattern frequency tracking with example collection
  * Pattern clustering for strategic insights
  * Missing capability identification
  * Severity scoring (logarithmic frequency + critical keywords)
  * Cluster priority (Critical/High/Medium/Low)

- ArchitectureEvolver for autonomous system evolution
  * Evolution triggering based on failure rate thresholds
  * Creates specialized agents for detected patterns

- SpecializedAgentFactory for dynamic agent creation
  * LLM-generated agent specifications from failure patterns
  * Custom system prompts and capability definitions

API Endpoints:
- POST /dev/metacognitive/analyze-performance
- POST /dev/metacognitive/analyze-failures

Testing:
- 12 comprehensive unit tests for all metacognitive components
- Tests cover degradation detection, bottleneck identification, trends

Bug Fixes:
- Fixed type mismatch in quantum_mcts.rs (n_visits u32 → usize)
- Fixed moved value error in analyzer.rs (added .clone())
- Fixed borrowing issues in arena.rs single-elimination tournament

Version: 0.5.0 → 0.6.0
Roadmap: Week 5-7 Complete ✅

+++ crates/beagle-agents/src/quantum_mcts.rs
@@ -106,7 +106,7 @@ impl QuantumMCTS {
             let metadata = HypothesisMetadata {
                 source: "mcts".to_string(),
                 confidence: hyp.q_value,
-                evidence_count: hyp.n_visits,
+                evidence_count: hyp.n_visits as usize,
                 created_at: i as f64,
             };
